+++ demo/src/app/components/+dialog/dialog-section.component.ts
@@ -1,12 +1,13 @@
-import { Component, OnDestroy, TemplateRef } from '@angular/core';
+import { Component, OnDestroy, TemplateRef, Renderer2 } from '@angular/core';
 import { ThyDialog, ThyDialogConfig, ThyDialogSizes } from '../../../../../src/dialog';
-import { helpers } from '../../../../../src/util';
+import { helpers, keycodes } from '../../../../../src/util';
 import { DemoDialogContentComponent } from './dialog-content.component';
 import { Subject, of, defer } from 'rxjs';
 import { takeUntil, delay, map } from 'rxjs/operators';
 import { apiParameters } from './api-parameters';
 import { taskTypes } from '../+select/mock-data';
 import { DemoTreeSectionComponent } from '../+tree/tree-section.component';
+import { mixinUnsubscribe, MixinBase } from '../../../../../src/core';
 
 const exampleCode = `
 import { DialogContentComponent } from './dialog-content.component';
@@ -37,8 +38,8 @@ export class DialogComponent {
     selector: 'demo-dialog-section',
     templateUrl: './dialog-section.component.html'
 })
-export class DemoDialogSectionComponent implements OnDestroy {
-    private ngUnsubscribe$ = new Subject();
+export class DemoDialogSectionComponent extends mixinUnsubscribe(MixinBase) implements OnDestroy {
+    // private ngUnsubscribe$ = new Subject();
 
     public exampleCode: string = exampleCode;
 
@@ -55,11 +56,16 @@ export class DemoDialogSectionComponent implements OnDestroy {
 
     selectedItem = this.optionData[0];
 
+    hasShowDialog = false;
+
+    unsubscribe: () => void;
+
     public thyPrimaryAction = (event: Event) => {
         return of(true).pipe(delay(1000));
     };
 
-    constructor(public thyDialog: ThyDialog) {
+    constructor(public thyDialog: ThyDialog, private renderer: Renderer2) {
+        super();
         thyDialog
             .afterOpened()
             .pipe(takeUntil(this.ngUnsubscribe$))
@@ -67,6 +73,19 @@ export class DemoDialogSectionComponent implements OnDestroy {
                 console.log(dialog);
             });
         this.optionData = taskTypes;
+
+        this.unsubscribe = renderer.listen(document, 'keydown', (event: KeyboardEvent) => {
+            const isK = (event.ctrlKey || event.metaKey) && event.keyCode === keycodes.K;
+            if (!this.hasShowDialog && isK) {
+                this.openComponentDialog();
+            }
+        });
+        // document.addEventListener('keydown', (event: KeyboardEvent) => {
+        //     const isK = (event.ctrlKey || event.metaKey) && event.keyCode === keycodes.K;
+        //     if (!this.hasShowDialog && isK) {
+        //         this.openComponentDialog();
+        //     }
+        // });
     }
 
     openTemplateDialog(template: TemplateRef<any>) {
@@ -82,6 +101,7 @@ export class DemoDialogSectionComponent implements OnDestroy {
     }
 
     openComponentDialog() {
+        this.hasShowDialog = true;
         const dialogRef = this.thyDialog.open(
             DemoDialogContentComponent,
             Object.assign(
@@ -97,6 +117,7 @@ export class DemoDialogSectionComponent implements OnDestroy {
             console.log(`Dialog keydownEvents: ${event}`);
         });
         dialogRef.afterClosed().subscribe(result => {
+            this.hasShowDialog = false;
             console.log(`Dialog afterClosed result: ${result}`);
         });
     }
@@ -121,7 +142,9 @@ export class DemoDialogSectionComponent implements OnDestroy {
     }
 
     ngOnDestroy(): void {
-        this.ngUnsubscribe$.next();
-        this.ngUnsubscribe$.complete();
+        super.ngOnDestroy();
+        if (this.unsubscribe) {
+            this.unsubscribe();
+        }
     }
 }