+++ src/select/index.ts
@@ -1,6 +1,7 @@
 export * from './module';
-export * from './option-item.component';
+// export * from './option-item.component';
 export * from './option.component';
-export * from './select-container.component';
+export * from './option-group.component';
+// export * from './select-container.component';
 export { ThySelectCustomComponent } from './select-custom.component';
 export { ThySelectComponent } from './select.component';